+++ server/API/APIHandler.py
@@ -1,10 +1,21 @@
 
+import os
 import time
 import random
 import json
 import tornado.ioloop, tornado.web
 from tornado import gen
 from Modules.UnixSysInfos import *
+
+
+def liste_dump(folder):
+    dico = {}
+    for root, dirs, files in os.walk(folder):
+        for dump in files:
+            path = "./" + root + "/" + dump
+            dico[dump] = open(path).read().replace("login:", "").replace("password:", "").split("\n")[:-1]
+    return dico
+
 
 # Handler for ressources
 class APIHandler(tornado.web.RequestHandler):
@@ -65,6 +76,14 @@ class APIHandler(tornado.web.RequestHandler):
                     }
                 )
             )
-        elif path_request.startswith("dump"):
-            self.write("TODO")
+        elif path_request == "dump":
+            self.write(json.dumps(liste_dump("logs/dump")))
+        elif path_request.startswith("dump/"):
+            try:
+                if os.path.isfile("logs/" + path_request):
+                    self.write(open(path).read())
+                else:
+                    self.write("not found")
+            except:
+                self.write("error")
